pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "motor>=3.7.1",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.44.0",
    "xgboost>=3.2.0",
]
